pyproject.toml
[tool.poetry]
name = "drawer"
version = "0.1.0"
description = ""
authors = ["Gizzatov Amir <gizzatovamir777@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
plotly = "^5.24.1"
pandas = "1.4.2"
pyyaml = "^6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
